refactored: post content
