+++ lib/ui/widgets/export.dart
@@ -21,3 +21,6 @@ export 'post_action_add_reaction/export.dart';
 export 'post_action_comment/export.dart';
 export 'post_action_like/export.dart';
 export 'post_action_reaction/export.dart';
+export 'post_images_previewer/export.dart';
+export 'post_item_header/export.dart';
+export 'post_content_message/export.dart';
