+++ lib/ui/widgets/post_content/index.dart
@@ -1,11 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:mooncake/entities/entities.dart';
-// import 'package:mooncake/ui/ui.dart';
-
-// import 'header/export.dart';
-// import 'text/export.dart';
-// import 'images/export.dart';
-// import 'poll/export.dart';
+import 'package:mooncake/ui/ui.dart';
+import 'widgets/export.dart';
 
 /// Contains the main content of a post. Such content is made of
 /// - The header of the post, indicating the creator and the data
