+++ lib/ui/widgets/post_content/widgets/post_poll_content/widgets/post_poll_result_item/index.dart
