+++ lib/ui/widgets/post_item_header/widgets/post_more_button/widgets/dialogs/widgets/popup_report/widgets/popup_report_text_input/index.dart
@@ -1,0 +1,30 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:mooncake/ui/ui.dart';
+import '../../../../blocs/export.dart';
+
+/// Represents the input that is used by the user to write something about
+/// the report.
+class PopupReportTextInput extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return BlocBuilder<ReportPopupBloc, ReportPopupState>(
+      builder: (context, state) {
+        return Padding(
+          padding: EdgeInsets.symmetric(horizontal: 16),
+          child: TextFormField(
+            autofocus: false,
+            onChanged: (value) {
+              BlocProvider.of<ReportPopupBloc>(context)
+                  .add(ChangeOtherText(value));
+            },
+            decoration: InputDecoration(
+              hintText: PostsLocalizations.of(context).reportPopupEditBotHint,
+              border: OutlineInputBorder(),
+            ),
+          ),
+        );
+      },
+    );
+  }
+}
